+++ src/main/java/com/magento/idea/magento2plugin/actions/generation/dialog/NewBlockDialog.java
@@ -112,11 +112,8 @@ public class NewBlockDialog extends AbstractDialog {
     }
 
     protected void onWriteActionOK() {
-
-        if (validateFormFields()) {
-            generateFile();
-            exit();
-        }
+        generateFile();
+        exit();
     }
 
     private void generateFile() {