+++ src/main/java/com/magento/idea/magento2plugin/actions/generation/dialog/NewCronjobDialog.java
@@ -290,10 +290,6 @@ public class NewCronjobDialog extends AbstractDialog {
      * When new cronjob dialog is filled, validate the input data and generate a new cronjob.
      */
     protected void onWriteActionOK() {
-
-        if (!validateFormFields()) {
-            return;
-        }
         final NamespaceBuilder namespaceBuilder = new NamespaceBuilder(
                 this.getCronjobModule(),
                 this.getCronjobClassName(),