+++ src/main/java/com/magento/idea/magento2plugin/actions/generation/dialog/NewGraphQlResolverDialog.java
@@ -122,11 +122,8 @@ public class NewGraphQlResolverDialog extends AbstractDialog {
     }
 
     protected void onWriteActionOK() {
-
-        if (validateFormFields()) {
-            generateFile();
-            exit();
-        }
+        generateFile();
+        exit();
     }
 
     private void generateFile() {