+++ src/main/java/com/magento/idea/magento2plugin/actions/generation/dialog/NewModuleDialog.java
@@ -179,11 +179,7 @@ public class NewModuleDialog extends AbstractDialog implements ListSelectionList
     }
 
     protected void onWriteActionOK() {
-
-        if (validateFormFields()) {
-            generateFiles();
-        }
-        exit();
+        generateFiles();
     }
 
     private void generateFiles() {