+++ src/main/java/com/magento/idea/magento2plugin/actions/generation/dialog/NewUiComponentFormDialog.java
@@ -437,16 +437,14 @@ public class NewUiComponentFormDialog extends AbstractDialog {
             fields.getCellEditor().stopCellEditing();
         }
 
-        if (validateFormFields()) {
-            generateRoutesXmlFile();
-            generateViewControllerFile();
-            generateSubmitControllerFile();
-            generateDataProviderFile();
-            generateLayoutFile();
-            generateFormFile();
-            generateAclXmlFile();
-            exit();
-        }
+        generateRoutesXmlFile();
+        generateViewControllerFile();
+        generateSubmitControllerFile();
+        generateDataProviderFile();
+        generateLayoutFile();
+        generateFormFile();
+        generateAclXmlFile();
+        exit();
     }
 
     /**