+++ src/main/java/com/magento/idea/magento2plugin/actions/generation/dialog/eavattribute/EavAttributeDialog.java
@@ -256,15 +256,8 @@ public abstract class  EavAttributeDialog extends AbstractDialog {
     }
 
     protected void onWriteActionOK() {
-
         stopOptionsTableEditing(getOptionsTable());
-
-        if (!validateFormFields()) {
-            return;
-        }
-
         generateExtraFilesBeforeDataPatchGeneration();
-
         final EavEntityDataInterface eavEntityDataInterface = getEavEntityData();
         generateDataPatchFile(eavEntityDataInterface);
         generateExtraFilesAfterDataPatchGeneration(eavEntityDataInterface);