+++ src/main/java/com/magento/idea/magento2plugin/indexes/ModuleIndex.java
@@ -7,23 +7,19 @@ package com.magento.idea.magento2plugin.indexes;
 
 import com.intellij.openapi.project.DumbService;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.vfs.VfsUtilCore;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiDirectory;
 import com.intellij.psi.PsiManager;
 import com.intellij.psi.search.GlobalSearchScope;
 import com.intellij.util.indexing.FileBasedIndex;
 import com.jetbrains.php.lang.PhpFileType;
-import com.magento.idea.magento2plugin.magento.packages.Package;
-import com.magento.idea.magento2plugin.project.util.GetProjectBasePath;
 import com.magento.idea.magento2plugin.stubs.indexes.ModuleNameIndex;
 import com.magento.idea.magento2plugin.util.RegExUtil;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import com.magento.idea.magento2plugin.util.magento.IsFileInEditableModuleUtil;
 import org.jetbrains.annotations.Nullable;
 
 public final class ModuleIndex {
@@ -40,41 +36,38 @@ public final class ModuleIndex {
     }
 
     public List<String> getEditableModuleNames() {
-        return getModuleNames(Package.vendor, true);
+        return getModuleNames(true);
     }
 
     public List<String> getEditableThemeNames() {
-        return getThemeNames("/" + Package.vendor + "/magento/|/tests/|/test/", true);
+        return getThemeNames(true);
     }
 
     public List<String> getModuleNames() {
-        return getModuleNames("/tests/|/test/", false);
+        return getModuleNames(false);
     }
 
     /**
      * Returns Module Names.
      *
-     * @param filterPattern String
      * @param withinProject boolean
      * @return List
      */
-    public List<String> getModuleNames(final String filterPattern, final boolean withinProject) {
-        return getNames(filterPattern, withinProject, RegExUtil.Magento.MODULE_NAME);
+    public List<String> getModuleNames(final boolean withinProject) {
+        return getNames(withinProject, RegExUtil.Magento.MODULE_NAME);
     }
 
     /**
      * Returns Theme Names.
      *
-     * @param filterPattern String
      * @param withinProject boolean
      * @return List
      */
-    public List<String> getThemeNames(final String filterPattern, final boolean withinProject) {
-        return getNames(filterPattern, withinProject, RegExUtil.Magento.THEME_NAME);
+    public List<String> getThemeNames(final boolean withinProject) {
+        return getNames(withinProject, RegExUtil.Magento.THEME_NAME);
     }
 
     private List<String> getNames(
-            final String filterPattern,
             final boolean withinProject,
             final String pattern
     ) {
@@ -82,7 +75,6 @@ public final class ModuleIndex {
                 .getInstance();
         final List<String> allModulesList = new ArrayList<>();
         final Collection<String> allModules = index.getAllKeys(ModuleNameIndex.KEY, project);
-        final Pattern compiled = Pattern.compile(filterPattern);
         for (final String moduleName : allModules) {
             if (!moduleName.matches(pattern)) {
                 continue;
@@ -96,17 +88,14 @@ public final class ModuleIndex {
             if (files.isEmpty()) {
                 continue;
             }
-            final VirtualFile virtualFile = files.iterator().next();
-            if (withinProject && !VfsUtilCore
-                    .isAncestor(GetProjectBasePath.execute(project), virtualFile, false)) {
-                continue;
-            }
+            for (final VirtualFile virtualFile : files) {
+                if (withinProject && !IsFileInEditableModuleUtil.execute(project, virtualFile)) {
+                    continue;
+                }
 
-            final Matcher matcher = compiled.matcher(virtualFile.getPath());
-            if (matcher.find()) {
-                continue;
+                allModulesList.add(moduleName);
+                break;
             }
-            allModulesList.add(moduleName);
         }
         Collections.sort(allModulesList);
         return allModulesList;